requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
bitarray==2.9.2
certifi==2024.6.2
charset-normalizer==3.3.2
ckzg==1.0.2
cytoolz==0.12.3
decorator==4.4.2
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.3.1
eth-utils==4.1.1
eth_abi==5.1.0
exceptiongroup==1.2.1
executing==2.0.1
frozenlist==1.4.1
hexbytes==0.3.1
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.5.1
ipython==8.25.0
jedi==0.19.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
matplotlib-inline==0.1.7
moviepy==1.0.3
multidict==6.0.5
numpy==2.0.0
opencv-python==4.10.0.84
parsimonious==0.10.0
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
proglog==0.1.10
prompt_toolkit==3.0.47
protobuf==5.27.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodome==3.20.0
Pygments==2.18.0
pyunormalize==15.1.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rlp==4.0.1
rpds-py==0.18.1
six==1.16.0
stack-data==0.6.3
toolz==0.12.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
web3==6.19.0
websockets==12.0
yarl==1.9.4
